requirements/base.txt
Babel==1.3
Django==1.8
Jinja2==2.7.3
Markdown==2.6.1
MarkupSafe==0.23
Pillow==2.7.0
PyYAML==3.11
Pygments==2.0.2
Sphinx==1.3.1
Unidecode==0.04.17
Werkzeug==0.10.1
alabaster==0.7.3
coverage==3.7.1
cssselect==0.9.1
dj-database-url==0.3.0
django-autoslug==1.7.2
django-braces==1.4.0
django-cache-url==0.8.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
django-extensions==1.5.1
django-filter==0.9.2
django-floppyforms==1.3.0
django-jinja==1.3.2
django-model-utils==2.2
django-rest-swagger==0.2.9
django-secure==1.0.1
djangorestframework==3.1.1
docutils==0.12
gnureadline==6.3.3
ipython==3.0.0
keen==0.3.12
lxml==3.4.2
psycopg2==2.6
pyquery==1.2.9
pytz==2014.10
raven==5.2.0
requests==2.6.0
six==1.9.0
snowballstemmer==1.2.0
sphinx-rtd-theme==0.1.7
sqlparse==0.1.14
unicode-slugify==0.1.3
wsgiref==0.1.2
